删除不存在的地址时不抛出异常

+++ src/Server.php
@@ -247,6 +247,9 @@ class Server
         $success = true;
 
         foreach ($paths as $path) {
+
+            if (!$this->exists($path)) continue;
+
             try {
                 if ($this->isTempFile($path)) {
                     if (!$this->destroy($path)) {
